fix: move all routes into handlers, create docs test
