+++ src/handlers/documentation.ts
@@ -1,0 +1,26 @@
+
+const docsHtml = `
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>CityCoins API Documentation</title>
+</head>
+<body style="margin: 0; padding: 0;">
+<redoc spec-url='http://petstore.swagger.io/v2/swagger.json'></redoc>
+<script src="https://cdn.jsdelivr.net/npm/redoc@latest/bundles/redoc.standalone.js"> </script>
+</body>
+</html>
+`
+
+const Documentation = async (): Promise<Response> => {
+  // return response
+  const headers = {
+    'Access-Control-Allow-Origin': '*',
+    'Content-Type': 'text/html; charset=utf-8',
+  }
+  return new Response(docsHtml, { headers })
+}
+
+export default Documentation
